requirements.txt
pandas
numpy
scikit-learn
imblearn
xgboost
matplotlib
seaborn
streamlit
joblib
tensorflow-cpu
streamlit